requirements.txt
scikit-learn
pdfplumber
python-docx
pandas
numpy
streamlit